+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 ## [Unreleased]
+### Added
+- `Object.null`
+- REPL history, with up/down keys choosing previous/next line
+
+### Fixed
+- Properties with `nil` as default value not accesible from Lua
+
 
 ## [r1]
 ### Added